requirements.txt
fastapi>=0.110
uvicorn>=0.27
pydantic>=2.0
requests>=2.31
pyyaml>=6.0
